Personal Site on Github
Github lets people create websites *freely hosted* without paying. It's hosted under *username*.github.io.
This can be changed, likely by paying a price, or taking the content and just hosting it elsewhere.

However, it's pretty solid to have this account.

MIT offered similar free hosting (not sure if we can still access those - but that ISA site I made was based on that).

Note that Github may 'lag' changes so you may have to wait a bit for the deployment to project your changes.

Bug: N/A (my personal site)
Test: TBD, will test once live